Use RAPTOR_BAD_CAST and RAPTOR_GOOD_CAST
Some good uses:
- narrowing a known, checked unicode char to a U16

Some 'bad' uses:
- only handing error messages, literal language, qname prefixes of a
  max len constrained by int
- passing in data to libxml constrained by int max len
- locator column field constrained to int size

Some bad uses:
- iostream read_bytes and write-bytes methods return int but could
  easily return a lot more in the size_t range (compare to fread).
  API change needed.
- locator byte field constrained to int size.  should be size_t
- raptor_nfc_check returns int offset into a buffer that could be
  larger

raptor_ntriples_parser_context changed line_length and offset to
size_t

raptor_turtle_parser changed buffer_length to size_t

+++ src/turtle_common.h
@@ -46,7 +46,7 @@ struct raptor_turtle_parser_s {
   char *buffer;
 
   /* buffer length */
-  int buffer_length;
+  size_t buffer_length;
   
   /* static statement for use in passing to user code */
   raptor_statement statement;
